+++ src/easy.rs
@@ -397,8 +397,8 @@ impl Config {
     // ///
     // /// If `force` is `true`, this matches the way cargo's `--config` CLI option
     // /// overrides config.
-    // pub fn merge(&mut self, from: Self, force: bool) -> Result<()> {
-    //     merge::Merge::merge(self, from, force)
+    // pub fn merge(&mut self, low: Self, force: bool) -> Result<()> {
+    //     merge::Merge::merge(self, low, force)
     // }
 }